requirements.txt
# Core dependencies
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Technical indicators
ta>=0.11.0
pandas-ta>=0.3.14b

# Async support
aiohttp>=3.9.0
websockets>=12.0

# Data validation
pydantic>=2.5.0

# Logging
structlog>=24.1.0
